Restructure project to WXT layout
