+++ manifest.json
@@ -15,7 +15,7 @@
     "<all_urls>"
   ],
   "background": {
-    "service_worker": "js/background.js",
+    "service_worker": "entrypoints/background.js",
     "type": "module"
   },
   "content_scripts": [
@@ -24,21 +24,21 @@
         "<all_urls>"
       ],
       "js": [
-        "js/content.js"
+        "entrypoints/content.js"
       ],
       "run_at": "document_start",
       "all_frames": false
     }
   ],
   "action": {
-    "default_popup": "popup/popup.html",
+    "default_popup": "entrypoints/popup/index.html",
     "default_icon": {
       "16": "icons/icon16.png",
       "48": "icons/icon48.png",
       "128": "icons/icon128.png"
     }
   },
-  "options_page": "options/options.html",
+  "options_page": "entrypoints/options/index.html",
   "icons": {
     "16": "icons/icon16.png",
     "48": "icons/icon48.png",
