+++ src/components/StatsTab.js
@@ -1,6 +1,6 @@
 import { h } from 'preact';
-import htm from './../js/lib/htm.mjs';
-import { getMessage } from './../js/modules/i18n.js';
+import htm from '../utils/htm.mjs';
+import { getMessage } from '../modules/i18n.js';
 
 const html = htm.bind(h);
