+++ src/entrypoints/options/index.html
@@ -7,6 +7,6 @@
     <link rel="stylesheet" href="../css/options.css">
 </head>
 <body>
-    <div id="options-app"></div> <script type="module" src="options.js"></script>
+    <div id="options-app"></div> <script type="module" src="index.js"></script>
 </body>
 </html>
