+++ src/entrypoints/popup/index.html
@@ -6,6 +6,6 @@
     <link rel="stylesheet" href="../css/popup.css">
 </head>
 <body>
-    <div id="popup-app"></div> <script type="module" src="popup.js"></script>
+    <div id="popup-app"></div> <script type="module" src="index.js"></script>
 </body>
 </html>
