+++ src/modules/i18n.js
